Add files via upload
This repository contains a machine learning projects, each focused on a different classification problem.
2. Email Classification
In this project, the focus is on classifying emails as either spam or non-spam (ham). The notebook walks through data preprocessing, feature extraction using techniques like TF-IDF, and the development of a machine learning model to accurately categorize the emails.